+++ test/modules/automation/protection-rule/dto/mocks/protection-rule.dto.mock.ts
