+++ test/modules/automation/protection-rule/params/protection-rule.params.mock.ts
